+++ Language/Namers/LambdaNamer.cpp
@@ -58,7 +58,7 @@ String LambdaNamer::defineBodyHtml() const
 {
 	if (!body())
 		return String::null;
-	return String(L"<div id=\"%1-body\" ondblclick=\"event.stopPropagation()\" style=\"display: none\">").arg((int)self()) + toHtml(body()) + L"</div>";
+	return String(L"<div id=\"%1-body\" ondblclick=\"event.stopPropagation()\" style=\"display: none\">").arg((long)self()) + toHtml(body()) + L"</div>";
 }
 
 String LambdaNamer::defineMidHtml(String const& _middle) const
@@ -71,7 +71,7 @@ String LambdaNamer::defineMidHtml(String const& _middle) const
 		else
 			info = " (empty)";
 	}
-	return _middle + tagOf(L"", tagOf(L"minor symbol", L";") + tagOf(L"_ms_minor", info), String(L"span id=\"%1-info\"").arg((int)self()));
+	return _middle + tagOf(L"", tagOf(L"minor symbol", L";") + tagOf(L"_ms_minor", info), String(L"span id=\"%1-info\"").arg((long)self()));
 }
 
 String LambdaNamer::defineEnclosureHtml(String const& _part, String const& _middle) const
@@ -81,7 +81,7 @@ String LambdaNamer::defineEnclosureHtml(String const& _part, String const& _midd
 
 String LambdaNamer::defineLambdaHtml(String const& _middle) const
 {
-	return String(L"<div onKeyPress=\"if (event=='{' ? set2('%1-info', '%1-body') : event=='}' ? set1('%1-info', '%1-body') : false) { CodeView.markDirty(%1); return true; } return false;\" onDblClick=\"toggle('%1-info', '%1-body'); CodeView.markDirty(%1); event.stopPropagation();\">").arg((int)self())
+	return String(L"<div onKeyPress=\"if (event=='{' ? set2('%1-info', '%1-body') : event=='}' ? set1('%1-info', '%1-body') : false) { CodeView.markDirty(%1); return true; } return false;\" onDblClick=\"toggle('%1-info', '%1-body'); CodeView.markDirty(%1); event.stopPropagation();\">").arg((long)self())
 		+ definePreHtml()
 		+ defineEnclosureHtml(L"head", defineReturnHtml() + defineNameHtml() + defineArgListHtml() + defineMidHtml(_middle))
 		+ defineEnclosureHtml(L"body", defineBodyHtml()) + definePostHtml() + L"</div>";
